Moved StringHelper, having one internal method left
