+++ FileHelpers/Attributes/ConditionalRecordAttribute.cs
@@ -1,5 +1,4 @@
 using System;
-using FileHelpers.Helpers;
 
 namespace FileHelpers
 {
