+++ FileHelpers/Core/ForwardReader.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Text;
-using FileHelpers.Helpers;
 
 namespace FileHelpers
 {
