+++ FileHelpers/Core/LineInfo.cs
@@ -2,7 +2,6 @@ using System;
 using System.Diagnostics;
 using System.Globalization;
 using FileHelpers.Core;
-using FileHelpers.Helpers;
 
 namespace FileHelpers
 {
