+++ FileHelpers/Dynamic/CsvClassBuilder.cs
@@ -2,7 +2,6 @@
 using System;
 using System.IO;
 using System.Text;
-using FileHelpers.Helpers;
 using FileHelpers.Options;
 
 namespace FileHelpers.Dynamic
