+++ FileHelpers/Engines/FileHelperAsyncEngine.cs
@@ -6,7 +6,6 @@ using System.IO;
 using System.Text;
 using FileHelpers.Engines;
 using FileHelpers.Events;
-using FileHelpers.Helpers;
 using FileHelpers.Streams;
 
 namespace FileHelpers
