+++ FileHelpers/Engines/FileHelperEngine.cs
@@ -8,7 +8,6 @@ using System.IO;
 using System.Text;
 using FileHelpers.Engines;
 using FileHelpers.Events;
-using FileHelpers.Helpers;
 using FileHelpers.Streams;
 
 namespace FileHelpers
