+++ FileHelpers/ErrorHandling/ErrorInfo.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Diagnostics;
-using FileHelpers.Helpers;
 
 namespace FileHelpers
 {
