+++ FileHelpers/ErrorHandling/ErrorManager.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
-using FileHelpers.Helpers;
 
 namespace FileHelpers
 {
