+++ FileHelpers/Fields/DelimitedField.cs
@@ -3,7 +3,6 @@ using System.Globalization;
 using System.Reflection;
 using System.Text;
 using FileHelpers.Core;
-using FileHelpers.Helpers;
 
 namespace FileHelpers
 {
