+++ FileHelpers/Fields/FixedLengthField.cs
@@ -3,7 +3,6 @@ using System.Reflection;
 using System.Text;
 using FileHelpers.Enums;
 using FileHelpers.Fields;
-using FileHelpers.Helpers;
 
 namespace FileHelpers
 {
