+++ FileHelpers/FormatDetector/SmartFormatDetector.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.Text;
 using FileHelpers.Dynamic;
 using System.IO;
-using FileHelpers.Helpers;
 
 namespace FileHelpers.Detection
 {
